Revert "Refactoring and added tslint."
This reverts commit 8ff39de61b5d0bd1e5ea091186fb0fb6a539e5f6.
