--- Something/WebApplication/wwwroot/js/global.scss
@@ -1,15 +1,0 @@
-body, .ant-form, h1, h2, h3, h4, h5, p, .ant-btn {
-    font-family: 'Lato', 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif !important;
-}
-
-.ovrd-btn.ant-btn {
-    background: #2f54eb;
-    border: 1px solid #2f54eb;
-    padding: 0 24px;
-    border-radius: 100px;
-    font-size: 16px;
-    line-height: 38px;
-    display: inline-block;
-    height: 40px;
-    text-align: center;
-}
