--- Something/WebApplication/wwwroot/js/index.tsx
@@ -1,9 +1,0 @@
-﻿import * as React from 'react';
-import * as ReactDOM from 'react-dom';
-
-import Message from './Message';
-
-ReactDOM.render(
-    <Message />,
-    document.getElementById('root')
-);
