+++ Something/WebApplication/wwwroot/js/login-app/components/Splash.tsx
@@ -1,7 +1,5 @@
 import * as React from 'react';
 
-import './Splash.scss';
-
 export class Splash extends React.Component {
     render() {
         return (
